Moved response parsing to a separate class and added JSON normalization
