+++ allms/models/abstract.py
@@ -13,7 +13,6 @@ from langchain.chains import LLMChain
 from langchain.chat_models.base import BaseChatModel
 from langchain.output_parsers import PydanticOutputParser
 from langchain.prompts import ChatPromptTemplate
-from langchain.schema import OutputParserException
 from langchain_core.language_models.llms import create_base_retry_decorator
 from langchain_core.prompts import HumanMessagePromptTemplate, SystemMessagePromptTemplate
 from langchain_core.prompts.prompt import PromptTemplate
@@ -32,8 +31,9 @@ from allms.defaults.long_text_chain import LongTextChainDefaults
 from allms.domain.enumerables import AggregationLogicForLongInputData, LanguageModelTask
 from allms.domain.input_data import InputData
 from allms.domain.prompt_dto import SummaryOutputClass, KeywordsOutputClass
-from allms.domain.response import ResponseWithError, ResponseData
+from allms.domain.response import ResponseData
 from allms.utils.long_text_processing_utils import get_max_allowed_number_of_tokens
+from allms.utils.response_parsing_utils import ResponseParser
 
 logger = logging.getLogger(__name__)
 
@@ -104,59 +104,8 @@ class AbstractModel(ABC):
         )
 
         if output_data_model_class:
-            return self._parse_model_output(model_responses)
+            return ResponseParser(self._parser).parse_model_output(model_responses)
         return model_responses
-    
-    def _extract_json_from_response(self, model_response_data: ResponseData) -> str:
-        search_results = self._json_pattern.findall(model_response_data.response)
-        
-        if len(search_results) == 0:
-            return model_response_data.response
-        
-        return search_results[0]
-
-    def _parse_response(
-            self, 
-            model_response_data: ResponseData
-        ) -> ResponseWithError:
-        raw_response = self._extract_json_from_response(model_response_data)
-
-        try:
-            return ResponseWithError(
-                response=self._parser.parse(raw_response), 
-                error_message=None
-            )
-        except OutputParserException as output_parser_exception:
-            return ResponseWithError(
-                response=None, 
-                error_message=f"""
-                    An OutputParserException has occurred for the model response: {raw_response}
-                    The exception message: {output_parser_exception}
-                    """
-            )
-
-    def _parse_model_output(
-            self, 
-            model_responses_data: typing.List[ResponseData]
-        ) -> typing.List[ResponseData]:
-        parsed_responses = []
-
-        for model_response_data in model_responses_data:
-            if not model_response_data.error:
-                response_with_error = self._parse_response(model_response_data)
-
-                parsed_responses.append(ResponseData(
-                    input_data=model_response_data.input_data,
-                    response=response_with_error.response,
-                    error=response_with_error.error_message,
-                    number_of_prompt_tokens=model_response_data.number_of_prompt_tokens,
-                    number_of_generated_tokens=model_response_data.number_of_generated_tokens
-
-                ))
-            else:
-                parsed_responses.append(model_response_data)
-
-        return parsed_responses
 
     async def _generate(
             self,
